Fix normal calcs
Former-commit-id: 0acc8d5c11efb66c3c98d84095e1ae42afcad52a

+++ src/engine/matrix.h
@@ -88,9 +88,9 @@ struct MatrixOps {
 };
 
 struct Color {
-  unsigned r;
-  unsigned g;
-  unsigned b;
+  double r;
+  double g;
+  double b;
 };
 
 #endif // MATRIX_H
